# Conflicts:
#	sormas-api/src/main/java/de/symeda/sormas/api/contact/ContactDto.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/common/DatabaseHelper.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/contact/edit/ContactEditFragment.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/contact/ContactFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/contact/ContactService.java
#	sormas-backend/src/test/java/de/symeda/sormas/backend/AbstractBeanTest.java

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/campaign/CampaignController.java
@@ -210,7 +210,7 @@ public class CampaignController {
 			component.addDiscardListener(discardCallback::run);
 		}
 
-		if (showDeleteButton) {
+		if (showDeleteButton && UserProvider.getCurrent().hasUserRight(UserRight.CAMPAIGN_DELETE)) {
 			component.addDeleteListener(() -> {
 				FacadeProvider.getCampaignFormDataFacade().deleteCampaignFormData(campaignFormDataUuid);
 				UI.getCurrent().getNavigator().navigateTo(CampaignFormDataView.VIEW_NAME);
